moved component files
...to src folder for easy access
